+++ src/diff-snapshot.js
@@ -206,7 +206,7 @@ function diffImageToSnapshot(options) {
   let result = {};
   const baselineSnapshotPath = path.join(snapshotsDir, `${snapshotIdentifier}-snap.png`);
   if (!fs.existsSync(baselineSnapshotPath)) {
-    mkdirp.sync(snapshotsDir);
+    mkdirp.sync(path.dirname(baselineSnapshotPath));
     fs.writeFileSync(baselineSnapshotPath, receivedImageBuffer);
     result = { added: true };
   } else {
@@ -270,7 +270,7 @@ function diffImageToSnapshot(options) {
     });
 
     if (isFailure({ pass, updateSnapshot })) {
-      mkdirp.sync(diffDir);
+      mkdirp.sync(path.dirname(diffOutputPath));
       const composer = new ImageComposer({
         direction: diffDirection,
       });
@@ -308,7 +308,7 @@ function diffImageToSnapshot(options) {
         imgSrcString: `data:image/png;base64,${pngBuffer.toString('base64')}`,
       };
     } else if (shouldUpdate({ pass, updateSnapshot, updatePassedSnapshot })) {
-      mkdirp.sync(snapshotsDir);
+      mkdirp.sync(path.dirname(baselineSnapshotPath));
       fs.writeFileSync(baselineSnapshotPath, receivedImageBuffer);
       result = { updated: true };
     } else {
@@ -359,7 +359,7 @@ function runDiffImageToSnapshot(options) {
     const output = writeDiffProcess.output[3].toString();
     result = JSON.parse(output);
   } else {
-    throw new Error('Error running image diff.');
+    throw new Error(`Error running image diff: ${(writeDiffProcess.error && writeDiffProcess.error.message) || 'Unknown Error'}`);
   }
 
   return result;